requirements.txt
streamlit
langchain
langgraph
langchain-groq
chromadb
sentence-transformers
pypdf
python-dotenv
uuid